// packages/companion-module-base/src/host-api/api.ts
export type EncodeIsVisible<T extends CompanionInputFieldBase> = Omit<T, 'isVisible' | 'isVisibleExpression'> & {
	isVisibleFn?: string
	isVisibleFnType?: 'function' | 'expression'
}
// ...
export interface SetFeedbackDefinitionsMessage {
	feedbacks: Array<{
		id: string
		name: string
		description: string | undefined
		options: EncodeIsVisible<SomeCompanionFeedbackInputField>[] // TODO module-lib - versioned types?
		type: 'boolean' | 'value' | 'advanced'
		defaultStyle?: CompanionFeedbackButtonStyleResult
		hasLearn: boolean
		showInvert: boolean | undefined
		learnTimeout: number | undefined
	}>
}
// ...
export interface UpdateFeedbackValuesMessage {
	values: Array<{
		id: string
		controlId: string
		value: JsonValue | Partial<CompanionFeedbackButtonStyleResult> | undefined
	}>
}

export interface FeedbackInstanceBase {
	id: string

	// If this is pending being run through upgrade scripts, the version it needs upgraded from is tracked here
	upgradeIndex: number | null
	disabled: boolean

	feedbackId: string // aka 'type'
	options: OptionsObject
}

export interface FeedbackInstance extends FeedbackInstanceBase {
	controlId: string

	isInverted: boolean

	/** If control supports an imageBuffer, the dimensions the buffer must be */
	image?: {
		width: number
		height: number
	}
}
// ...
export interface ParseVariablesInStringMessage {
	text: string
	controlId: string | undefined
	feedbackInstanceId: string | undefined
	actionInstanceId: string | undefined
}
export interface ParseVariablesInStringResponseMessage {
	text: string
	variableIds: string[] | undefined
}